+++ Chungwoo/CloneBaemin/app/src/main/java/com/lcw/study/clonebaemin/feature/favorite/FavoriteFragment.kt
@@ -1,10 +1,11 @@
-package com.lcw.study.clonebaemin
+package com.lcw.study.clonebaemin.feature.favorite
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.lcw.study.clonebaemin.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
