+++ Chungwoo/CloneBaemin/app/src/main/java/com/lcw/study/clonebaemin/feature/main/MainActivity.kt
@@ -1,9 +1,15 @@
-package com.lcw.study.clonebaemin
+package com.lcw.study.clonebaemin.feature.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import com.lcw.study.clonebaemin.R
 import com.lcw.study.clonebaemin.databinding.ActivityMainBinding
+import com.lcw.study.clonebaemin.feature.favorite.FavoriteFragment
+import com.lcw.study.clonebaemin.feature.home.HomeFragment
+import com.lcw.study.clonebaemin.feature.myinfo.MyifoFragment
+import com.lcw.study.clonebaemin.feature.orderlist.OrderListFragment
+import com.lcw.study.clonebaemin.feature.search.SearchFragment
 
 class MainActivity : AppCompatActivity() {
     private var mBinding: ActivityMainBinding? = null
