+++ Chungwoo/CloneBaemin/app/src/main/res/layout/activity_splash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashActivity">
+    tools:context=".feature.splash.SplashActivity">
 
     <TextView
         android:layout_width="match_parent"
